backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
tensorflow==2.16.1
tf-keras>=2.16.0
numpy==1.26.3
pillow>=10.2.0
pydantic>=2.6.1
pydantic-core>=2.16.2
python-dotenv>=1.0.1
aiofiles>=23.2.1
